feat: add `withdraw` to Factory
To minimize the number of contracts for which a browser user will need a
FunctionCall access key, this adds a `withdraw` function to
BridgeTokenFactory and an associated `withdraw_internal` to BridgeToken.
`withdraw_internal` follows the pattern of similar BridgeToken
functions, and is only callable by the BridgeTokenFactory that deployed
this BridgeToken.